Layout and planning (#1)
* _sidebar changes

* Updated the sidebar to show the planned layout. Would love some feedback.

* Didn't realize the nesting was automatic from the headings in the markdown files. I've reset the sidebar to follow the previous funtionality. Time to get to work on actual files.

* This will look better hopefully

* Layout Finalized

Fixed a couple typos and finalized the layout. I'm thinking of adding a navbar in addition to the side bar, but that'll be in a separate branch

* Rename 0002_UnOfficial_Mythras_Starter_Set.md to 0002_Unofficial_Mythras_Starter_Set.md

* Update README.md

Fixed Unofficial Mythras Starter Set link

* Trying to fix the tables on Starter Set Page

* Reformated Tables on Starter Set Page

* Update index.html

* Title change and Background Colors

* Cover Delete

I didn't like the cover page. I got rid of it

* Update _sidebar.md

Filled in the TDM, DTRPG, and Discord links

* Update index.html

* Create _navbar.md

* Update index.html

ToC tweaks and navbar added fr... maybe

* Update _navbar.md

* Update _navbar.md

* Update _navbar.md

* Update index.html

* Update index.html

* Added WIP disclaimers

* Update index.html

The navbar was cringe on mobile @nd I don't feel like fixing it. I turned it off.

* Resources and Free Stuff

Filled out the resources and Free stuff pages

* Update 0003_Resources_for_Mythras.md

* Update 0003_Resources_for_Mythras.md

* Update 0003_Resources_for_Mythras.md

* Update 0003_Resources_for_Mythras.md

* Clean up

* SRD's Page

Populated the SRD page and clarified that the imperatives and the SRDs are one and the same.

* Basic Auto Calculating Character Sheet

* Update 0003_Resources_for_Mythras.md

* Mythic Britain Update

Experimenting with AI generated summaries since I'm not familiar with the products.

* Update 0006_Settings.md

* Settings And Direct Links

Finished settings though Thennla and added Direct links for all free material.

* Update 0006_Settings.md

* Thennla link fix

* Update 0006_Settings.md

* Update Thennla.pdf

* Update Mythras_Firearms.pdf

* Update 0006_Settings.md

Finished out the rest of the settings page

* Update 0006_Settings.md

* Update 0006_Settings.md

* Update 0007_Daughter_RPGs.md

* Renamed Daughter_RPGs

Renamed the daughter rpgs file to rulesets as well as updated all the references to it. Also deleted the Sidebar test

* Update _navbar.md

* Rules Supplements and Adventures

* Done?

I think that's everything

* Update 0009_Adventure_Supplements.md

* Update 0009_Adventure_Supplements.md

* Readability (Drop Downs)

* Update 0002_Unofficial_Mythras_Starter_Set.md

* Update 0009_Adventure_Supplements.md